requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aisuite=0.1.6=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.30.1=pypi_0
anyio=4.7.0=pypi_0
blas=1.0=mkl
blinker=1.9.0=pypi_0
brotli-python=1.0.9=py311hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.11.26=haa95532_0
certifi=2024.8.30=py311haa95532_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
cuda-cccl=12.6.77=0
cuda-cccl_win-64=12.6.77=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.6.77=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
distro=1.9.0=pypi_0
filelock=3.13.1=py311haa95532_0
flask=3.1.0=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2024.10.0=pypi_0
giflib=5.2.2=h7edc060_0
gmpy2=2.1.2=py311h7f96b67_0
groq=0.13.1=pypi_0
grpcio=1.68.1=pypi_0
grpcio-tools=1.68.1=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.26.5=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.7=py311haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=py311haa95532_1
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h827c3e9_3
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.7.77=0
libcurand-dev=10.3.7.77=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
lz4-c=1.9.4=h2bbff1b_1
markupsafe=2.1.3=py311h2bbff1b_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.11=py311h827c3e9_0
mkl_random=1.2.8=py311hea22821_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
networkx=3.2.1=py311haa95532_0
numpy=2.0.1=py311hdab7c0b_1
numpy-base=2.0.1=py311hd01c5d8_1
openai=1.57.4=pypi_0
openjpeg=2.5.2=hae555c5_0
openssl=3.0.15=h827c3e9_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pillow=11.0.0=py311hb5480e2_0
pip=24.2=py311haa95532_0
portalocker=2.10.1=pypi_0
protobuf=5.29.1=pypi_0
pydantic=2.10.3=pypi_0
pydantic-core=2.27.1=pypi_0
pypdf=5.1.0=pypi_0
pysocks=1.7.1=py311haa95532_0
python=3.11.11=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytorch=2.5.1=py3.11_cuda11.8_cudnn9_0
pytorch-cuda=11.8=h24eeafa_6
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pywin32=308=pypi_0
pyyaml=6.0.2=py311h827c3e9_0
qdrant-client=1.12.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=py311haa95532_1
safetensors=0.4.5=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=pypi_0
sentence-transformers=3.3.1=pypi_0
setuptools=75.1.0=py311haa95532_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tokenizers=0.21.0=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.47.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=py311haa95532_0
vc=14.40=haa95532_2
vs2015_runtime=14.42.34433=h9531ae6_2
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.6=h8880b57_0
